feat: Update templates and views for blog and home apps
This commit updates the templates and views for the blog and home apps. The 'blog.html' and 'home.html' templates are removed, and new 'index.html' templates are added for both apps. The views.py files are modified to render the 'index.html' templates instead of the deleted templates. Additionally, the 'base.html' template is added to the 'globals' directory, and the 'index.html' templates for both apps extend this base template.